--- src/main/java/com/wanted/preonboarding/ticket/global/exception/InvalidInputException.java
@@ -1,11 +1,0 @@
-package com.wanted.preonboarding.ticket.global.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.BAD_REQUEST)
-public class InvalidInputException extends RuntimeException {
-    public InvalidInputException(String message) {
-        super(message);
-    }
-}
