+++ src/main/java/com/wanted/preonboarding/ticket/presentation/controller/AlarmController.java
@@ -1,25 +1,13 @@
 package com.wanted.preonboarding.ticket.presentation.controller;
 
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.wanted.preonboarding.ticket.domain.dto.SmsResponse;
 import com.wanted.preonboarding.ticket.domain.dto.ReservePossibleAlarmCustomerInfoDto;
-import com.wanted.preonboarding.ticket.domain.dto.ResponseDto;
-import com.wanted.preonboarding.ticket.global.dto.BaseResDto;
-import com.wanted.preonboarding.ticket.global.exception.ResultCode;
-import com.wanted.preonboarding.ticket.global.exception.ServiceException;
+import com.wanted.preonboarding.ticket.aop.dto.BaseResDto;
 import com.wanted.preonboarding.ticket.presentation.service.AlarmSmsService;
-import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
-import java.io.UnsupportedEncodingException;
-import java.net.URISyntaxException;
-import java.security.InvalidKeyException;
-import java.security.NoSuchAlgorithmException;
 
 @Slf4j
 @ControllerAdvice
