+++ src/main/java/com/wanted/preonboarding/ticket/presentation/controller/QueryController.java
@@ -3,8 +3,7 @@ package com.wanted.preonboarding.ticket.presentation.controller;
 import com.wanted.preonboarding.core.domain.response.ResponseHandler;
 import com.wanted.preonboarding.ticket.application.TicketSeller;
 import com.wanted.preonboarding.ticket.domain.dto.*;
-import com.wanted.preonboarding.ticket.global.dto.BaseResDto;
-import com.wanted.preonboarding.ticket.global.exception.InvalidInputException;
+import com.wanted.preonboarding.ticket.aop.dto.BaseResDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
