+++ src/main/java/com/wanted/preonboarding/ticket/presentation/controller/ReserveController.java
@@ -1,27 +1,14 @@
 package com.wanted.preonboarding.ticket.presentation.controller;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.wanted.preonboarding.ticket.application.TicketSeller;
 import com.wanted.preonboarding.ticket.domain.dto.*;
-import com.wanted.preonboarding.ticket.global.dto.BaseResDto;
-import com.wanted.preonboarding.ticket.global.exception.InvalidInputException;
-import com.wanted.preonboarding.ticket.global.exception.ResultCode;
-import com.wanted.preonboarding.ticket.global.exception.ServiceException;
-import jakarta.persistence.EntityNotFoundException;
-import lombok.Builder;
+import com.wanted.preonboarding.ticket.aop.dto.BaseResDto;
+import com.wanted.preonboarding.ticket.aop.ResultCode;
+import com.wanted.preonboarding.ticket.aop.exception.ServiceException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.antlr.v4.runtime.misc.NotNull;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
-
-import javax.naming.NameNotFoundException;
-import java.io.UnsupportedEncodingException;
-import java.net.URISyntaxException;
-import java.security.InvalidKeyException;
-import java.security.NoSuchAlgorithmException;
 
 @Slf4j
 @RestController
